TAN-2049: refactor to create both components in the same file
