+++ packages/desktop/app/components/TriageTable.js
@@ -5,8 +5,7 @@ import { push } from 'connected-react-router';
 import { useEncounter } from '../contexts/Encounter';
 import { DataFetchingTable } from './Table';
 import { DateDisplay } from './DateDisplay';
-import { LocationGroupCell } from './LocationGroupCell';
-import { LocationCell } from './LocationCell';
+import { LocationCell, LocationGroupCell } from './LocationCell';
 import { TriageWaitTimeCell } from './TriageWaitTimeCell';
 import { useLocalisation } from '../contexts/Localisation';
 import { reloadPatient } from '../store';
