+++ packages/desktop/app/views/patients/PatientListingView.js
@@ -1,7 +1,6 @@
 import React, { useState, useCallback } from 'react';
 import { useDispatch } from 'react-redux';
-import { LocationCell } from '../../components/LocationCell';
-import { LocationGroupCell } from '../../components/LocationGroupCell';
+import { LocationCell, LocationGroupCell } from '../../components/LocationCell';
 import { usePatientNavigation } from '../../utils/usePatientNavigation';
 import { reloadPatient } from '../../store/patient';
